+++ src/PSRule/Pipeline/RulePipeline.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using PSRule.Rules;
 using System;
 using System.Management.Automation;