+++ CHANGELOG.md
@@ -13,6 +13,8 @@ What's changed since v0.21.0:
   - Updated `HasField` to check if any of the specified fields exist. [#578](https://github.com/microsoft/PSRule/issues/578)
 - General improvements:
   - Input format detection now includes `.jsonc` and `.markdown` file extensions. [#575](https://github.com/microsoft/PSRule/issues/575)
+  - Improved support for cross module rule dependencies. [#248](https://github.com/microsoft/PSRule/issues/248)
+    - Rule dependencies are now automatically imported.
 - Bug fixes:
   - Fixed handling for null or empty arrays with `StartsWith`, `Contains`, `EndsWith`, `In`, and `NotIn`. [#579](https://github.com/microsoft/PSRule/issues/579)